performed analysis on the 1st type and stats
conclusion there seems to be no given formula for what pokemon recevies what type according to stats, with the exception being bug types, it could be assumed this is because bug types usually are light and frail making them one of the few pokemon who's stats and type show a correlation.

Accuracy of range 20-30 could be imporved trough the introudction of more parameters.